SDFD-26: refactor fee stats component (#299)
* feat: fee stats integration

* feat: XML component

* wip: average transaction fee chart

* feat: average transaction fee chart

* feat: average transaction fee

* feat: average fee stats integration

* chore: remove mock file

* fix: NaN results

* feat: average transactions fee

* refactor: adjustments PR

* refactor: fee stats component

* refactor: average transaction fee component

* refactor: transaction per second component

* refactor: adjustments PR

* refactor: adjustments PR

* refactor: adjustments PR

* refactor: adjustments PR
