+++ frontend/src/components/AverageTransactionFee/index.tsx
@@ -86,8 +86,10 @@ export const AverageTransactionFee: React.FC = () => {
 
   return (
     <SectionCard
-      title="Average Transaction Fee"
+      title="Average Transaction Fee by Day (In Stroops):"
       titleCustom={timeRangeOptions}
+      titleLinkLabel="API"
+      titleLink="api/fees/stats"
       noData={!feeStats.fees}
     >
       <div className="AverageTransactionFee__mainChart">
