+++ frontend/src/components/FeeStats/index.tsx
@@ -1,11 +1,13 @@
-import React, { useCallback, useEffect } from "react";
-import { Icon, Table } from "@stellar/design-system";
+import React, { useCallback, useEffect, useMemo } from "react";
+import { Icon, Table, TextLink } from "@stellar/design-system";
 import { useDispatch } from "react-redux";
 import { get } from "lodash";
 import BigNumber from "bignumber.js";
 
 import { useRedux } from "hooks/useRedux";
 import { SectionCard } from "components/SectionCard";
+import { AmountInfoCard } from "components/AmountInfoCard";
+import { AverageTransactionFee } from "components/AverageTransactionFee";
 import { fetchFeeStatsDataAction } from "ducks/feeStats";
 import "./styles.scss";
 
@@ -92,8 +94,8 @@ export const FeeStats = () => {
 
       return (
         <React.Fragment key={data.id}>
-          <td className="FeeStats__metric">{data.name}</td>
-          <td className="FeeStats__value">
+          <td className="FeeStats__tableContainer__metric">{data.name}</td>
+          <td className="FeeStats__tableContainer__value">
             {feeResult}
             {isCapacityUsage ? capacityStyle(value) : iconStyle(value)}
           </td>
@@ -103,16 +105,53 @@ export const FeeStats = () => {
     [feeStats.data],
   );
 
+  const data = useMemo(() => {
+    if (feeStats.fees?.month) {
+      const daysInMonth = feeStats.fees.month.length;
+      const feesResults = feeStats.fees.month.reduce((acc, currentValue) => {
+        acc += Number(currentValue.primaryValue);
+        return acc;
+      }, 0);
+
+      const baseOperationFee = 0.00001;
+
+      const average = String((feesResults / daysInMonth) * baseOperationFee);
+
+      const formattedAverage = new BigNumber(average).toFormat(5);
+
+      return `${formattedAverage} XLM`;
+    }
+
+    return "0 XLM";
+  }, [feeStats.fees?.month]);
+
   return (
-    <SectionCard title="Fee Stats (Last 5 ledgers), Fee unit in stroops.">
+    <SectionCard title="Transaction Fee Info">
       <div className="FeeStats">
-        <Table
-          id="fee-stats"
-          data={nameMap}
-          columnLabels={labels}
-          hideNumberColumn
-          renderItemRow={renderFee}
-        />
+        <div className="FeeStats__cardsContainer">
+          <AmountInfoCard title="Average Transaction Fee" amount={data} />
+          <AmountInfoCard title="Base operation fee" amount="0.00001 XLM" />
+          <AmountInfoCard title="Base reserve" amount="0.5 XLM" />
+        </div>
+
+        <div className="FeeStats__tableContainer">
+          <div className="FeeStats__tableContainer__title">
+            Fee Stats (Last 5 ledgers), Fee unit in{" "}
+            <TextLink href="https://developers.stellar.org/docs/glossary/fees/#base-fee">
+              stroops
+            </TextLink>
+          </div>
+
+          <Table
+            id="fee-stats"
+            data={nameMap}
+            columnLabels={labels}
+            hideNumberColumn
+            renderItemRow={renderFee}
+          />
+        </div>
+
+        <AverageTransactionFee />
       </div>
     </SectionCard>
   );
