+++ frontend/src/components/FeeStats/styles.scss
@@ -1,13 +1,36 @@
 .FeeStats {
-  &__value {
+  &__cardsContainer {
     display: flex;
-    justify-content: flex-end;
     align-items: center;
+    flex-wrap: wrap;
+    gap: 1rem;
+  }
 
-    svg {
-      height: 0.9rem;
-      width: 0.9rem;
-      margin-left: 0.8rem;
+  &__tableContainer {
+    margin-bottom: 2rem;
+
+    &__title {
+      color: var(--pal-text-primary);
+      font-weight: var(--font-weight-medium);
+
+      text-transform: uppercase;
+      margin: 4rem 0 2rem;
+    }
+
+    &__value {
+      display: flex;
+      justify-content: flex-end;
+      align-items: center;
+
+      svg {
+        height: 0.9rem;
+        width: 0.9rem;
+        margin-left: 0.8rem;
+      }
+    }
+
+    &__link {
+      color: var(--pal-brand-primary);
     }
   }
 }
