--- frontend/src/components/XLM/styles.scss
@@ -1,6 +1,0 @@
-.XLMContainer {
-  display: flex;
-  align-items: center;
-  flex-wrap: wrap;
-  gap: 1rem;
-}
